requirements-dev.txt
-r requirements.txt

django-postgrespool==0.3.1
factory-boy==2.11.1
Faker==1.0.2
flake8==3.7.5
freezegun==0.3.11
PyHamcrest==1.9.0
pytest==4.2.1
pytest-django==3.4.7
pytest-factoryboy==2.0.2
pytest-lazy-fixture==0.5.1
pytest-mock==1.10.1
